Move UndoManager into EditableStyledDocument so that one UndoManager is shared across multiple clones. Clones can be disposed without affecting access to the UndoManager.
